pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-image-sourcing"
version = "0.1.0"
description = "Agent-facing image retrieval platform for public-web discovery, inspection, and persistence."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "fastapi>=0.124",
    "httpx>=0.28",
    "pillow>=12.0",
    "pydantic-settings>=2.12",
    "requests>=2.32",
    "uvicorn>=0.38",
]

[project.optional-dependencies]
mcp = ["mcp>=1.10"]
postgres = ["psycopg[binary]>=3.2"]
s3 = ["boto3>=1.35"]
test = ["pytest>=8.4"]

[project.scripts]
ais = "agentic_image_sourcing.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
